Command line options should be able to have more than one expected argument.
The command line options previously could only handle a single
option. As a bit of a hack the --jar-compare command was using a comma
separated value to accept two arguments.

This was causing issues around validation and path expansion. By separating
these arguments the shell should automatically handle expansions for things
like ~/ and $HOME for the second argument.

This fix should address issue #11 and issue #12.